Use u32 for LCG

+++ src/utils.rs
@@ -10,8 +10,8 @@ macro_rules! unwrap_or_return {
 
 /// Linear congruential generator
 struct Lcg {
-    m: usize,
-    state: usize,
+    m: u32,
+    state: u32,
 }
 
 impl Lcg {
@@ -21,18 +21,18 @@ impl Lcg {
             m,
             state: std::time::SystemTime::now()
                 .duration_since(std::time::UNIX_EPOCH)
-                .map_or_else(|_| 69696969, |duration| duration.as_secs() as usize)
+                .map_or_else(|_| 69696969, |duration| duration.as_secs() as u32)
                 % m,
         }
     }
 
     /// Int in range (max_val is exclusive)
-    fn rand_int_range(&mut self, max_val: usize) -> usize {
+    fn rand_int_range(&mut self, max_val: u32) -> u32 {
         let a = 48271;
         let c = 0;
 
         self.state = (a * self.state + c) % self.m;
-        ((self.state as f64 / self.m as f64) * max_val as f64) as usize
+        ((self.state as f64 / self.m as f64) * max_val as f64) as u32
     }
 }
 
@@ -41,7 +41,7 @@ pub fn shuffle<T>(vec: &mut Vec<T>) {
     let n = vec.len();
     let mut lcg = Lcg::new();
     for i in 0..n - 1 {
-        let j = i + lcg.rand_int_range(n) * (n - i) / n;
+        let j = i + lcg.rand_int_range(n as u32) as usize * (n - i) / n;
         vec.swap(i, j);
     }
 }
